move brave and tavily to remote
