+++ llama_stack/providers/registry/tool_runtime.py
@@ -19,14 +19,6 @@ def available_providers() -> List[ProviderSpec]:
     return [
         InlineProviderSpec(
             api=Api.tool_runtime,
-            provider_type="inline::brave-search",
-            pip_packages=[],
-            module="llama_stack.providers.inline.tool_runtime.brave_search",
-            config_class="llama_stack.providers.inline.tool_runtime.brave_search.config.BraveSearchToolConfig",
-            provider_data_validator="llama_stack.providers.inline.tool_runtime.brave_search.BraveSearchToolProviderDataValidator",
-        ),
-        InlineProviderSpec(
-            api=Api.tool_runtime,
             provider_type="inline::memory-runtime",
             pip_packages=[],
             module="llama_stack.providers.inline.tool_runtime.memory",
@@ -35,18 +27,30 @@ def available_providers() -> List[ProviderSpec]:
         ),
         InlineProviderSpec(
             api=Api.tool_runtime,
-            provider_type="inline::tavily-search",
-            pip_packages=[],
-            module="llama_stack.providers.inline.tool_runtime.tavily_search",
-            config_class="llama_stack.providers.inline.tool_runtime.tavily_search.config.TavilySearchToolConfig",
-            provider_data_validator="llama_stack.providers.inline.tool_runtime.tavily_search.TavilySearchToolProviderDataValidator",
-        ),
-        InlineProviderSpec(
-            api=Api.tool_runtime,
             provider_type="inline::code-interpreter",
             pip_packages=[],
             module="llama_stack.providers.inline.tool_runtime.code_interpreter",
             config_class="llama_stack.providers.inline.tool_runtime.code_interpreter.config.CodeInterpreterToolConfig",
+        ),
+        remote_provider_spec(
+            api=Api.tool_runtime,
+            adapter=AdapterSpec(
+                adapter_type="brave-search",
+                module="llama_stack.providers.remote.tool_runtime.brave_search",
+                config_class="llama_stack.providers.remote.tool_runtime.brave_search.config.BraveSearchToolConfig",
+                pip_packages=["requests"],
+                provider_data_validator="llama_stack.providers.remote.tool_runtime.brave_search.BraveSearchToolProviderDataValidator",
+            ),
+        ),
+        remote_provider_spec(
+            api=Api.tool_runtime,
+            adapter=AdapterSpec(
+                adapter_type="tavily-search",
+                module="llama_stack.providers.remote.tool_runtime.tavily_search",
+                config_class="llama_stack.providers.remote.tool_runtime.tavily_search.config.TavilySearchToolConfig",
+                pip_packages=["requests"],
+                provider_data_validator="llama_stack.providers.remote.tool_runtime.tavily_search.TavilySearchToolProviderDataValidator",
+            ),
         ),
         remote_provider_spec(
             api=Api.tool_runtime,
